pull isbn validation helpers into separate file
